+++ tests/conftest.py
@@ -1,0 +1,20 @@
+import pytest
+from app import create_app, db
+from app.controllers import init_db
+
+
+@pytest.fixture
+def client():
+    app = create_app(environment="testing")
+    app.config["TESTING"] = True
+
+    with app.test_client() as client:
+        app_ctx = app.app_context()
+        app_ctx.push()
+        db.drop_all()
+        db.create_all()
+        init_db(True)
+        yield client
+        db.session.remove()
+        db.drop_all()
+        app_ctx.pop()
