+++ tests/test_auth.py
@@ -1,25 +1,4 @@
-import pytest
-
-from app import db, create_app
-from app.controllers import init_db
 from tests.utils import login, logout
-
-
-@pytest.fixture
-def client():
-    app = create_app(environment="testing")
-    app.config["TESTING"] = True
-
-    with app.test_client() as client:
-        app_ctx = app.app_context()
-        app_ctx.push()
-        db.drop_all()
-        db.create_all()
-        init_db()
-        yield client
-        db.session.remove()
-        db.drop_all()
-        app_ctx.pop()
 
 
 def test_auth_pages(client):
