+++ tests/test_users.py
@@ -1,27 +1,7 @@
-import pytest
 from sqlalchemy import desc
 from flask.testing import FlaskClient
-from app import db, create_app
-from app.controllers import init_db
 from app.models import User
 from .utils import login
-
-
-@pytest.fixture
-def client():
-    app = create_app(environment="testing")
-    app.config["TESTING"] = True
-
-    with app.test_client() as client:
-        app_ctx = app.app_context()
-        app_ctx.push()
-        db.drop_all()
-        db.create_all()
-        init_db(True)
-        yield client
-        db.session.remove()
-        db.drop_all()
-        app_ctx.pop()
 
 
 def test_users_page(client: FlaskClient):
